software/Raspberry/requirements.txt
smbus2
python-decouple
requests
betterconf
dotenv
sqlalchemy